pyproject.toml
[project]
name = "nys-county-power"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apportionment==1.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "click==8.1.8",
    "comm==0.2.2",
    "contourpy==1.3.2",
    "cupy-cuda13x>=13.6.0",
    "cycler==0.12.1",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "executing==2.2.0",
    "fastrlock==0.8.3",
    "fonttools==4.57.0",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsonlines==4.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy==2.2.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyzmq==26.4.0",
    "seaborn==0.13.2",
    "six==1.17.0",
    "stack-data==0.6.3",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "tzdata==2025.2",
    "wcwidth==0.2.13",
]
